# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Avito Backend Test Task",
        "contact": {
            "name": "Artemiy Chaykov"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/balance/{id}": {
            "get": {
                "description": "get active balance of the user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User account"
                ],
                "summary": "balance",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/structures.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/book": {
            "put": {
                "description": "book money on user account before withdraw operation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User account"
                ],
                "summary": "book",
                "parameters": [
                    {
                        "description": "User id, money to book amount",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal.bookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/structures.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/deposit": {
            "put": {
                "description": "top up the user's balance by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User account"
                ],
                "summary": "deposit",
                "parameters": [
                    {
                        "description": "User from whom this money, User to whom this money, order id, service id, money amount",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal.transferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/structures.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/report": {
            "get": {
                "description": "create report CSV file with indicating the amount of revenue for each service, return url for the file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounting"
                ],
                "summary": "report",
                "parameters": [
                    {
                        "description": "month MM, year YYYY",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal.reportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/transactions": {
            "get": {
                "description": "get list of the transactions for user. User can set order by date or amount",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User account"
                ],
                "summary": "transactions",
                "parameters": [
                    {
                        "description": "user id, preferred order (date/amount/",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal.HistoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/structures.Order"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/unbook": {
            "put": {
                "description": "unbook money on user account if withdraw fall",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User account"
                ],
                "summary": "unbook",
                "parameters": [
                    {
                        "description": "User id, money to unbook amount",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal.bookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/structures.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/withdraw": {
            "put": {
                "description": "debiting money from a separate account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User account"
                ],
                "summary": "withdraw",
                "parameters": [
                    {
                        "description": "User from whom this money, User to whom this money, order id, service id, money amount",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal.transferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/structures.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "definitions": {
        "internal.HistoryRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "sortOrder": {
                    "type": "string"
                }
            }
        },
        "internal.bookRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "internal.reportRequest": {
            "type": "object",
            "properties": {
                "month": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "internal.transferRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "fromuserid": {
                    "type": "integer"
                },
                "orderid": {
                    "type": "integer"
                },
                "serviceid": {
                    "type": "integer"
                },
                "touserid": {
                    "type": "integer"
                }
            }
        },
        "structures.Order": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fromuserid": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "orderid": {
                    "type": "integer"
                },
                "serviceid": {
                    "type": "integer"
                },
                "touserid": {
                    "type": "integer"
                }
            }
        },
        "structures.User": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "bookamount": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                }
            }
        }
    }
}